refactor: move enam & update imports, make password field is optional, replace Error with BadRequestException
