+++ backend/src/auth/auth.enums.ts
@@ -1,0 +1,5 @@
+export enum OAuthProvider {
+  YANDEX = 'yandex',
+  GOOGLE = 'google',
+  GITHUB = 'github',
+}
